pydantic-ai/requirements.txt
pydantic
fastapi
uvicorn
pydantic-ai
openai
httpx
logfire
sqlalchemy
psycopg2-binary
langchain
langchain-community
supabase
PyPDF2
pandas
python-docx
pytesseract
streamlit
streamlit-chat
python-dotenv
opencv-python
langfuse
crawl4ai
docling
watchdog
langchain-ollama
numexpr
markdown
ddgs
